[#231] refactor: Resolution Widget 재활용 가능하도록 분리
