+++ lib/presentation/common_components/common_components.dart
@@ -1,5 +1,7 @@
 export 'colored_button.dart';
 export 'gradient_bottom_sheet.dart';
+export 'profile_image_circle.dart';
+export 'resolution_list_cell_widget.dart';
 export 'show_toast_message.dart';
 export 'user_profile_bar.dart';
 export 'wehavit_app_bar.dart';
