+++ lib/presentation/my_page/view/my_page_view_widget.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:fpdart/fpdart.dart';
 import 'package:loading_animation_widget/loading_animation_widget.dart';
 import 'package:wehavit/common/common.dart';
 import 'package:wehavit/dependency/domain/usecase_dependency.dart';
@@ -128,8 +127,7 @@ class MyProfileWidget extends StatelessWidget {
                     ),
                   ),
                   Text(
-                    // TODO: UserEmail 대신 UserMessage 보여주기
-                    userModel.userEmail ?? '',
+                    userModel.aboutMe ?? '',
                     style: const TextStyle(
                       fontSize: 14,
                       fontWeight: FontWeight.w300,
@@ -179,31 +177,6 @@ class MyProfileWidget extends StatelessWidget {
           ],
         );
       },
-    );
-  }
-}
-
-class ProfileImageCircleWidget extends StatelessWidget {
-  const ProfileImageCircleWidget({
-    required this.size,
-    required this.url,
-    super.key,
-  });
-
-  final String? url;
-  final double size;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: size,
-      height: size,
-      decoration: const BoxDecoration(shape: BoxShape.circle),
-      clipBehavior: Clip.hardEdge,
-      child: Image.network(
-        url ?? '',
-        fit: BoxFit.cover,
-      ),
     );
   }
 }
@@ -343,309 +316,6 @@ class SimpleStatisticsBulletWidget extends StatelessWidget {
           ),
         ),
       ],
-    );
-  }
-}
-
-class MyPageResolutionListCellWidget extends ConsumerStatefulWidget {
-  const MyPageResolutionListCellWidget({
-    super.key,
-    required this.resolutionEntity,
-    required this.showDetails,
-  });
-
-  final ResolutionEntity resolutionEntity;
-  final bool showDetails;
-
-  @override
-  ConsumerState<MyPageResolutionListCellWidget> createState() =>
-      _MyPageResolutionListCellWidgetState();
-}
-
-class _MyPageResolutionListCellWidgetState
-    extends ConsumerState<MyPageResolutionListCellWidget> {
-  @override
-  Widget build(BuildContext context) {
-    final EitherFuture<List<bool>> futureDoneList =
-        ref.watch(getTargetResolutionDoneListForWeekUsecaseProvider)(
-      resolutionId: widget.resolutionEntity.resolutionId ?? '',
-      startMonday: DateTime.now().getMondayDateTime(),
-    );
-
-    return Container(
-      margin: const EdgeInsets.only(bottom: 16.0),
-      width: double.infinity,
-      decoration: const BoxDecoration(
-        color: CustomColors.whSemiBlack,
-        borderRadius: BorderRadius.all(
-          Radius.circular(16.0),
-        ),
-      ),
-      child: Column(
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(
-              top: 20,
-              left: 20,
-              right: 20,
-              bottom: 20,
-            ),
-            child: Column(
-              children: [
-                ResolutionListCellHeadWidget(
-                  goalStatement: widget.resolutionEntity.goalStatement ?? '',
-                  resolutionName: widget.resolutionEntity.resolutionName ?? '',
-                  showGoalStatement: widget.showDetails,
-                  pointColor: PointColors
-                      .colorList[widget.resolutionEntity.colorIndex ?? 0],
-                ),
-                const SizedBox(
-                  height: 12,
-                ),
-                ResolutionLinearGaugeWidget(
-                  resolutionEntity: widget.resolutionEntity,
-                  futureDoneList: futureDoneList,
-                ),
-                SizedBox(
-                  height: widget.showDetails ? 20 : 4,
-                ),
-                Visibility(
-                  visible: widget.showDetails,
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Text(
-                        // ignore: lines_longer_than_80_chars
-                        '${DateTime.now().difference(DateTime.now().subtract(const Duration(days: 3))).inDays + 1}일째 도전 중',
-                        style: const TextStyle(
-                          color: CustomColors.whWhite,
-                          fontSize: 14.0,
-                          fontWeight: FontWeight.w500,
-                        ),
-                      ),
-                      ResolutionListWeeklyDoneWidget(
-                        futureDoneList: futureDoneList,
-                        pointColor: PointColors
-                            .colorList[widget.resolutionEntity.colorIndex ?? 0],
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class ResolutionListCellHeadWidget extends StatelessWidget {
-  const ResolutionListCellHeadWidget({
-    super.key,
-    required this.resolutionName,
-    required this.goalStatement,
-    required this.showGoalStatement,
-    required this.pointColor,
-  });
-
-  final String resolutionName;
-  final String goalStatement;
-  final bool showGoalStatement;
-  final Color pointColor;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Container(
-          width: 36,
-          height: 36,
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(8.0),
-            color: pointColor,
-          ),
-        ),
-        const SizedBox(
-          width: 12,
-        ),
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                resolutionName,
-                style: TextStyle(
-                  color: pointColor,
-                  fontSize: 16.0,
-                  fontWeight: FontWeight.w600,
-                  height: 1.2,
-                ),
-              ),
-              Visibility(
-                visible: showGoalStatement,
-                child: Text(
-                  goalStatement,
-                  style: const TextStyle(
-                    color: CustomColors.whWhite,
-                    fontSize: 14.0,
-                    fontWeight: FontWeight.w500,
-                    height: 1.2,
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-        const SizedBox(
-          width: 12,
-        ),
-        Icon(
-          Icons.chevron_right,
-          size: 32,
-          color: pointColor,
-        ),
-      ],
-    );
-  }
-}
-
-class ResolutionListWeeklyDoneWidget extends StatelessWidget {
-  const ResolutionListWeeklyDoneWidget({
-    super.key,
-    required this.futureDoneList,
-    required this.pointColor,
-  });
-
-  final EitherFuture<List<bool>> futureDoneList;
-  final Color pointColor;
-
-  @override
-  Widget build(BuildContext context) {
-    return EitherFutureBuilder<List<bool>>(
-      target: futureDoneList,
-      forWaiting: Row(
-        children: List<Widget>.generate(
-          7,
-          (index) => Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 2.0),
-            child: ResolutionListWeeklyDoneCellPlaceholderWidget(),
-          ),
-        ),
-      ),
-      forFail: Row(
-        children: List<Widget>.generate(
-          7,
-          (index) => Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 2.0),
-            child: ResolutionListWeeklyDoneCellPlaceholderWidget(),
-          ),
-        ),
-      ),
-      mainWidgetCallback: (doneList) {
-        return Row(
-          children: List<Widget>.generate(
-            7,
-            (index) => Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 2.0),
-              child: ResolutionListWeeklyDoneCellWidget(
-                isDone: doneList[index],
-                weekday: index,
-                pointColor: pointColor,
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
-///
-/// weekday는 월요일부터 0 ~ 일요일 6
-class ResolutionListWeeklyDoneCellWidget extends StatelessWidget {
-  const ResolutionListWeeklyDoneCellWidget({
-    super.key,
-    required this.isDone,
-    required this.weekday,
-    required this.pointColor,
-  });
-
-  final int weekday;
-  final bool isDone;
-  final Color pointColor;
-
-  @override
-  Widget build(BuildContext context) {
-    return Visibility(
-      visible: isDone,
-      replacement: Container(
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          color: CustomColors.whSemiBlack,
-          border: Border.all(
-            color: CustomColors.whBrightGrey,
-            width: 2,
-          ),
-        ),
-        width: 25,
-        height: 25,
-        alignment: Alignment.center,
-        child: Text(
-          weekdayKorean[weekday],
-          style: const TextStyle(
-            color: CustomColors.whBrightGrey,
-            fontSize: 13.0,
-            fontWeight: FontWeight.w600,
-          ),
-        ),
-      ),
-      child: Container(
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          color: pointColor,
-        ),
-        alignment: Alignment.center,
-        width: 25,
-        height: 25,
-        child: const Icon(
-          Icons.check,
-          color: CustomColors.whWhite,
-          size: 22,
-          weight: 100,
-        ),
-      ),
-    );
-  }
-}
-
-class ResolutionListWeeklyDoneCellPlaceholderWidget extends StatelessWidget {
-  const ResolutionListWeeklyDoneCellPlaceholderWidget({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        shape: BoxShape.circle,
-        color: CustomColors.whSemiBlack,
-        border: Border.all(
-          color: CustomColors.whBrightGrey,
-          width: 2,
-        ),
-      ),
-      width: 25,
-      height: 25,
-      alignment: Alignment.center,
-      child: Container(
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          color: CustomColors.whBrightGrey,
-        ),
-        width: 4,
-        height: 4,
-      ),
     );
   }
 }
